refactor(models, scripts): job-storeをjob-store-apiへ移動・importパスを整理
- job-store配下をjob-store-api配下へリネーム・移動
- ルートindex.tsや型定義のエクスポート先を新構成に変更
- scripts/copy-schema.tsのdrizzleスキーマ参照パスを新構成に対応
- その他importパスの修正

BREAKING CHANGE: job-store関連のimportパスが変更されました。利用側は新しいパスに修正が必要です。
