+++ packages/models/src/job-store-api/type.ts
@@ -1,9 +1,9 @@
 import type { InferOutput } from "valibot";
-import type { cursorSchema, decodedNextTokenSchema } from "./jobList/continue";
-import type { JobListSchema, JobSchema, searchFilterSchema } from "./client";
+import type { JobListSchema, JobSchema, searchFilterSchema } from "./dbClient";
 import type { jobs, jobSelectSchema } from "./drizzle";
-import type { insertJobRequestBodySchema } from "./jobInsert";
-import type { jobListQuerySchema } from "./jobList";
+import type { cursorSchema, decodedNextTokenSchema } from "./endpoints/jobListContinue";
+import type { insertJobRequestBodySchema } from "./endpoints/jobInsert";
+import type { jobListQuerySchema } from "./endpoints/jobList";
 
 export type Cursor = InferOutput<typeof cursorSchema>;
 // --- コマンド型 ---
@@ -62,8 +62,8 @@ export type CommandOutput<T extends JobStoreCommand> = T extends {
   type: infer U;
 }
   ? U extends keyof CommandOutputMap
-    ? CommandOutputMap[U]
-    : never
+  ? CommandOutputMap[U]
+  : never
   : never;
 
 // --- コマンドパターンなDBクライアント ---
@@ -74,8 +74,8 @@ export type JobStoreDBClient = {
 // 🔍 型チェック用ユーティリティ
 export type KeysMustMatch<A, B> = Exclude<keyof A, keyof B> extends never
   ? Exclude<keyof B, keyof A> extends never
-    ? true
-    : ["Extra keys in B:", Exclude<keyof B, keyof A>]
+  ? true
+  : ["Extra keys in B:", Exclude<keyof B, keyof A>]
   : ["Extra keys in A:", Exclude<keyof A, keyof B>];
 
 type JobSelectFromDrizzle = typeof jobs.$inferSelect;
